MajorChange#2
-Have taken away single lane roads and replaced them with two lane roads
Used a struct for each grid rather than new header files, each struct contains 2 pointers
made printgrids alot better

going to begin working on save/load , just wanted this code backed up

+++ EE273-Traffic-Sim/SimulationEngine.cpp
@@ -132,8 +132,8 @@ void SimulationEngine::load(std::string filename) {
 	}
 	std::string line;
 	(getline(inFile, line)); // Is not in a loop as we only want the first line to be used for the purpose of setting the time
-		setCurrenttime(stod(line));
-		setInitialtime(stod(line));
+		setCurrenttime(stoi(line));
+		setInitialtime(stoi(line));
 
 		inFile.close();
